# docs/swagger.yaml
basePath: /
definitions:
  entity.Comment:
    properties:
      createdAt:
        type: string
      id:
        type: integer
      message:
        type: string
      photo_id:
        type: integer
      updatedAt:
        type: string
      user_id:
        type: integer
    type: object
  entity.CommentRequest:
    properties:
      message:
        type: string
      photo_id:
        type: integer
    type: object
...
  helper.BadRequest:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
  helper.InternalServerError:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
  helper.SuccessResult:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: febrianto.bekasi@gmail.com
    name: febrianto
  description: Final Project Digitalent x Hactive8
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Mygram
  version: "1.0"
paths:
  /comment:
    get:
      consumes:
      - application/json
      description: Get details of all comment
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/helper.SuccessResult'
            - properties:
                code:
                  type: integer
                data:
                  $ref: '#/definitions/entity.Comment'
                message:
                  type: string
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/helper.BadRequest'
            - properties:
                code:
                  type: integer
                message:
                  type: string
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/helper.InternalServerError'
            - properties:
                code:
                  type: integer
                message:
                  type: string
              type: object
      security:
      - JWT: []
      summary: Get all comment
      tags:
      - comment
    post:
      consumes:
      - application/json
      description: Create new comment
      parameters:
      - description: create comment
        in: body
        name: comment
        required: true
        schema:
          $ref: '#/definitions/entity.CommentRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/helper.SuccessResult'
            - properties:
                code:
                  type: integer
                data:
                  $ref: '#/definitions/entity.Comment'
                message:
                  type: string
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/helper.BadRequest'
            - properties:
                code:
                  type: integer
                message:
                  type: string
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/helper.InternalServerError'
            - properties:
                code:
                  type: integer
                message:
                  type: string
              type: object
      security:
      - JWT: []
      summary: Create comment
      tags:
      - comment
  /comment/{id}:
    delete:
      consumes:
      - application/json
      description: Delete the comment corresponding to the input Id
      parameters:
      - description: ID of the comment to be deleted
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/helper.SuccessResult'
            - properties:
                code:
                  type: integer
                data:
                  $ref: '#/definitions/entity.Comment'
                message:
                  type: string
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/helper.BadRequest'
            - properties:
                code:
                  type: integer
                message:
                  type: string
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/helper.InternalServerError'
            - properties:
                code:
                  type: integer
                message:
                  type: string
              type: object
      security:
      - JWT: []
      summary: Delete comment identified by the given id
      tags:
      - comment
    get:
      consumes:
      - application/json
      description: Get details of comment corresponding is the input Id
      parameters:
      - description: ID of the comment
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/helper.SuccessResult'
            - properties:
                code:
                  type: integer
                data:
                  $ref: '#/definitions/entity.Comment'
                message:
                  type: string
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/helper.BadRequest'
            - properties:
                code:
                  type: integer
                message:
                  type: string
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/helper.InternalServerError'
            - properties:
                code:
                  type: integer
                message:
                  type: string
              type: object
      security:
      - JWT: []
      summary: Get details for a given id
      tags:
      - comment
    put:
      consumes:
      - application/json
      description: Update the comment corresponding to the input id
      parameters:
      - description: create comment
        in: body
        name: comment
        required: true
        schema:
          $ref: '#/definitions/entity.CommentRequest'
      - description: ID of the comment
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/helper.SuccessResult'
            - properties:
                code:
                  type: integer
                data:
                  $ref: '#/definitions/entity.Comment'
                message:
                  type: string
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/helper.BadRequest'
            - properties:
                code:
                  type: integer
                message:
                  type: string
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/helper.InternalServerError'
            - properties:
                code:
                  type: integer
                message:
                  type: string
              type: object
      security:
      - JWT: []
      summary: Update comment identified by the given id
      tags:
      - comment
...
schemes:
- http
securityDefinitions:
  JWT:
    description: 'How to input in swagger : ''Bearer <insert_your_token_here>'''
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"